backend/requirements.txt
flask
flask-cors
pymongo
python-dotenv
PyJWT
werkzeug
numpy
scipy
matplotlib
pandas
